Added explicit definition for np.random.randint

+++ pimms/random_number.pyx
@@ -19,5 +19,5 @@ def seed_randint_np(int seedval):
 
 
 def generate_random_numbers(int min_val, int max_val, int size):        
-    cdef cnp.ndarray[cnp.int_t, ndim=1] random_numbers = np.random.randint(min_val, max_val, size, dtype=np.int32)
+    cdef cnp.ndarray[cnp.int_t, ndim=1] random_numbers = np.random.randint(min_val, max_val, size, dtype=np.int64)
     return np.asarray(random_numbers)
